+++ src/__tests__/settings.test.js
@@ -11,23 +11,31 @@
 
 import EventEmitter from 'events';
 import ProjectWorkspace from '../project_workspace';
-import Settings from '../Settings';
+import getSettings from '../Settings';
 
-describe('Settings', () => {
-  it('sets itself up fom the constructor', () => {
-    const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'test', 1000);
-    const options = {
-      shell: true,
-    };
-    const settings = new Settings(workspace, options);
-    expect(settings.workspace).toEqual(workspace);
-    expect(settings.settings).toEqual(expect.any(Object));
-    expect(settings.spawnOptions).toEqual(options);
-  });
+function prepareProcess() {
+  const mockProcess: any = new EventEmitter();
+  mockProcess.stdout = new EventEmitter();
+  mockProcess.stderr = new EventEmitter();
 
-  it('[jest 20] reads and parses the config', () => {
+  return {
+    createProcess: jest.fn(() => mockProcess),
+    mockProcessResult: (stdout: ?string, stderr: ?string) => {
+      if (stdout) {
+        mockProcess.stdout.emit('data', Buffer.from(stdout));
+      }
+      if (stderr) {
+        mockProcess.stderr.emit('data', Buffer.from(stderr));
+      }
+      mockProcess.emit('close');
+    },
+  };
+}
+
+describe('getSettings', () => {
+  it('[jest 19] reads and parses the config', async () => {
+    expect.assertions(2);
     const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'test', 1000);
-    const completed = jest.fn();
     const config = {
       cacheDirectory: '/tmp/jest',
       name: '[md5 hash]',
@@ -37,26 +45,20 @@ describe('Settings', () => {
       version: '19.0.0',
     };
 
-    const mockProcess: any = new EventEmitter();
-    mockProcess.stdout = new EventEmitter();
-    const createProcess = () => mockProcess;
-    const buffer = makeBuffer(JSON.stringify(json));
-    const settings = new Settings(workspace, {
+    const {createProcess, mockProcessResult} = prepareProcess();
+    const settingsPromise = getSettings(workspace, {
       createProcess,
     });
+    mockProcessResult(JSON.stringify(json));
+    const settings = await settingsPromise;
 
-    settings.getConfig(completed);
-    settings.getConfigProcess.stdout.emit('data', buffer);
-    settings.getConfigProcess.emit('close');
-
-    expect(completed).toHaveBeenCalled();
     expect(settings.jestVersionMajor).toBe(19);
-    expect(settings.settings).toEqual(config);
+    expect(settings.configs).toEqual([config]);
   });
 
-  it('[jest 21] reads and parses the config', () => {
+  it('[jest 21] reads and parses the config', async () => {
+    expect.assertions(2);
     const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'test', 1000);
-    const completed = jest.fn();
     const configs = [
       {
         cacheDirectory: '/tmp/jest',
@@ -68,69 +70,59 @@ describe('Settings', () => {
       version: '21.0.0',
     };
 
-    const mockProcess: any = new EventEmitter();
-    mockProcess.stdout = new EventEmitter();
-    const createProcess = () => mockProcess;
-    const buffer = makeBuffer(JSON.stringify(json));
-    const settings = new Settings(workspace, {
+    const {createProcess, mockProcessResult} = prepareProcess();
+    const settingsPromise = getSettings(workspace, {
       createProcess,
     });
+    mockProcessResult(JSON.stringify(json));
+    const settings = await settingsPromise;
 
-    settings.getConfig(completed);
-    settings.getConfigProcess.stdout.emit('data', buffer);
-    settings.getConfigProcess.emit('close');
-
-    expect(completed).toHaveBeenCalled();
-    expect(settings.jestVersionMajor).toBe(21);
-    expect(settings.settings).toEqual(configs[0]);
-  });
-
-  it('[jest 21] reads and parses the configs', () => {
-    const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'test', 1000);
-    const completed = jest.fn();
-    const configs = [
-      {
-        cacheDirectory: '/tmp/jest',
-        name: '[md5 hash]',
-      },
-    ];
-    const json = {
-      configs,
-      version: '21.0.0',
-    };
-
-    const mockProcess: any = new EventEmitter();
-    mockProcess.stdout = new EventEmitter();
-    const createProcess = () => mockProcess;
-    const buffer = makeBuffer(JSON.stringify(json));
-    const settings = new Settings(workspace, {
-      createProcess,
-    });
-
-    settings.getConfigs(completed);
-    settings.getConfigProcess.stdout.emit('data', buffer);
-    settings.getConfigProcess.emit('close');
-
-    expect(completed).toHaveBeenCalled();
     expect(settings.jestVersionMajor).toBe(21);
     expect(settings.configs).toEqual(configs);
   });
 
-  it('calls callback even if no data is sent', () => {
+  it('rejects the promise even if no data is sent', async () => {
+    expect.assertions(1);
     const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'test', 1000);
-    const completed = jest.fn();
 
-    const mockProcess: any = new EventEmitter();
-    mockProcess.stdout = new EventEmitter();
-    const createProcess = () => mockProcess;
-    const settings = new Settings(workspace, {
+    const {createProcess, mockProcessResult} = prepareProcess();
+    const settingsPromise = getSettings(workspace, {
       createProcess,
     });
+    mockProcessResult();
 
-    settings.getConfig(completed);
-    settings.getConfigProcess.emit('close');
+    try {
+      await settingsPromise;
+    } catch (err) {
+      expect(err).toBeTruthy();
+    }
+  });
 
-    expect(completed).toHaveBeenCalled();
+  it('rejects the promise if an error is logged', async () => {
+    expect.assertions(1);
+    const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'test', 1000);
+    const configs = [
+      {
+        cacheDirectory: '/tmp/jest',
+        name: '[md5 hash]',
+      },
+    ];
+    const json = {
+      configs,
+      version: '21.0.0',
+    };
+
+    const {createProcess, mockProcessResult} = prepareProcess();
+    const settingsPromise = getSettings(workspace, {
+      createProcess,
+    });
+    mockProcessResult(JSON.stringify(json), 'The expected error occured.');
+
+    try {
+      await settingsPromise;
+    } catch (err) {
+      expect(err).toBeTruthy();
+    }
   });
 
   it('passes command, args, and options to createProcess', () => {
@@ -138,19 +130,13 @@ describe('Settings', () => {
     const pathToConfig = 'test';
     const pathToJest = 'path_to_jest';
     const rootPath = 'root_path';
-
     const workspace = new ProjectWorkspace(rootPath, pathToJest, pathToConfig, localJestMajorVersion);
-    const createProcess = jest.fn().mockReturnValue({
-      on: () => {},
-      stdout: new EventEmitter(),
-    });
 
-    const options: any = {
+    const {createProcess} = prepareProcess();
+    getSettings(workspace, {
       createProcess,
       shell: true,
-    };
-    const settings = new Settings(workspace, options);
-    settings.getConfig(() => {});
+    });
 
     expect(createProcess).toBeCalledWith(
       {
@@ -168,7 +154,6 @@ describe('Settings', () => {
 
   describe('parse config', () => {
     const workspace = new ProjectWorkspace('root_path', 'path_to_jest', 'test', 1000);
-    const createProcess = jest.fn();
 
     const json = `{ 
       "version": "23.2.0",
@@ -176,37 +161,23 @@ describe('Settings', () => {
         "testRegex": "some-regex"
       }]
     }`;
-    const run_test = (text: string, expected_version: number = 23, expected_regex: string = 'some-regex'): void => {
-      settings._parseConfig(text);
-      const target = settings.configs[0];
-      expect(settings.jestVersionMajor).toBe(expected_version);
-      expect(target.testRegex).toBe(expected_regex);
-    };
-
-    let settings;
-    beforeEach(() => {
-      settings = new Settings(workspace, {
+    const run_test = async (
+      text: string,
+      expected_version: number = 23,
+      expected_regex: string = 'some-regex'
+    ): Promise<void> => {
+      expect.assertions(2);
+      const {createProcess, mockProcessResult} = prepareProcess();
+      const settingsPromise = getSettings(workspace, {
         createProcess,
       });
-    });
+      mockProcessResult(text);
+      const settings = await settingsPromise;
 
-    it('test regex', () => {
-      const regex = settings._jsonPattern;
+      expect(settings.jestVersionMajor).toBe(expected_version);
+      expect(settings.configs[0].testRegex).toBe(expected_regex);
+    };
 
-      let text = ` > abc {}
-        { abc }
-      `;
-      let index = text.search(regex);
-      expect(index).not.toBe(-1);
-      expect(text.substring(index).trim()).toBe('{ abc }');
-
-      text = `{def: 
-        {sub}
-      }`;
-      index = text.search(regex);
-      expect(index).not.toBe(-1);
-      expect(text.substring(index).startsWith('{def:')).toBe(true);
-    });
     it('can parse correct config', () => {
       run_test(json);
     });
@@ -221,12 +192,3 @@ describe('Settings', () => {
     });
   });
 });
-
-const makeBuffer = (content: string) => {
-  // Buffer.from is not supported in < Node 5.10
-  if (typeof Buffer.from === 'function') {
-    return Buffer.from(content);
-  }
-
-  return Buffer.from(content);
-};