+++ src/index.js
@@ -11,7 +11,7 @@ import * as Process from './Process';
 
 import ProjectWorkspace from './project_workspace';
 import Runner from './Runner';
-import Settings from './Settings';
+import getSettings from './Settings';
 import Snapshot from './Snapshot';
 import {
   Expect,
@@ -39,7 +39,7 @@ module.exports = {
   Process,
   ProjectWorkspace,
   Runner,
-  Settings,
+  getSettings,
   Snapshot,
   TestReconciler,
   parse,